+ResponseEntity<> в контроллерах +@ControllerAdvice +@ExceptionHandler

+++ src/main/java/com/example/crossFit/service/RequestFitService.java
@@ -1,8 +1,8 @@
 package com.example.crossFit.service;
 
 
-import com.example.crossFit.exceptions.EntityAlreadyIsRegisteredException;
-import com.example.crossFit.exceptions.EntityNotFoundException;
+import com.example.crossFit.exceptions.ResourceAlreadyIsRegisteredException;
+import com.example.crossFit.exceptions.ResourceNotFoundException;
 import com.example.crossFit.model.entity.Accountant;
 import com.example.crossFit.model.entity.Client;
 import com.example.crossFit.model.entity.RequestFit;
@@ -10,7 +10,6 @@ import com.example.crossFit.repository.ClientRepo;
 import com.example.crossFit.repository.RequestFitRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpStatus;
 import org.springframework.mail.SimpleMailMessage;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -51,13 +50,15 @@ public class RequestFitService {
 
     @PreAuthorize("hasRole('ROLE_USER')")
     @Transactional
-    public void createRequest(RequestFit requestFit) {
+    public String createRequest(RequestFit requestFit) {
         RequestFit requestFit1 = requestFitRepo.findByPhoneNumber(requestFit.getPhoneNumber());
         if (requestFit1 != null) {
-            throw new EntityAlreadyIsRegisteredException(HttpStatus.BAD_REQUEST,
-                    "Заявка с таким номером уже создана");
+            throw new ResourceAlreadyIsRegisteredException("Заявка с указанным номером телефона "
+                    + requestFit.getPhoneNumber() + " уже была зарегистрирована!");
         }
         requestFitRepo.save(requestFit);
+
+        return "Заявка успешно зарегистрирована!";
     }
 
     @PreAuthorize("hasRole('ROLE_ADMIN')")
@@ -65,8 +66,8 @@ public class RequestFitService {
     public RequestFit findByPhoneNumber(String phoneNumber) {
         RequestFit requestFit = requestFitRepo.findByPhoneNumber(phoneNumber);
         if (requestFit == null) {
-            throw new EntityNotFoundException(HttpStatus.NOT_FOUND,
-                    "Заявка с указанным номером телефона не найдена");
+            throw new ResourceNotFoundException("Заявка с указанным номером телефона:" + phoneNumber
+                    + " не найдена!");
         }
         return requestFit;
     }
@@ -76,53 +77,62 @@ public class RequestFitService {
     public List<RequestFit> showAllRequestFitIsNotApprove() {
         List<RequestFit> requestFitsList = requestFitRepo.findAllByIsApprovedNull();
         if (requestFitsList.isEmpty()) {
-            throw new EntityNotFoundException(HttpStatus.NOT_FOUND,
-                    "Новых заявок на абонемент не найдено");
+            throw new ResourceNotFoundException("Новых заявок на абонемент не найдено!");
         }
         return requestFitsList;
     }
 
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    @PreAuthorize("hasRole('ROLE_ADMIN') or hasRole('ROLE_USER')")
     @Transactional(readOnly = true)
     public List<RequestFit> findAllRequestFit() {
+        if (requestFitRepo.findAll().isEmpty()) {
+            throw new ResourceNotFoundException("Заявок не найдено!");
+        }
         return requestFitRepo.findAll();
     }
 
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     @Transactional
-    public void deleteRequestFit(String phoneNumber) {
-        findByPhoneNumber(phoneNumber);
+    public String deleteRequestFit(String phoneNumber) {
+        if (requestFitRepo.findByPhoneNumber(phoneNumber) == null) {
+            throw new ResourceNotFoundException("Заявка с указанным номером телефона: " + phoneNumber
+                    + " не зарегистрирована!");
+        }
         requestFitRepo.deleteByPhoneNumber(phoneNumber);
+
+        return "Заявка успешно удалена";
     }
 
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     @Transactional
-    public void rejectRequestFit(String phoneNumber) {
+    public String rejectRequestFit(String phoneNumber) {
         RequestFit requestFit = requestFitRepo.findByPhoneNumber(phoneNumber);
         if (requestFit == null) {
-            throw new EntityNotFoundException(HttpStatus.NOT_FOUND,
-                    "Заявка с указанным номером телефона не найдена");
+            throw new ResourceNotFoundException("Заявка с указанным номером телефона: "
+                    + phoneNumber + " не найдена!");
         }
         requestFit.setApproved(false);
 
         requestFitRepo.save(requestFit);
 
         sendMessage(requestFit.getEmail(), "Отказ", "К сожалению, Ваша заявка отклонена!");
+
+        return "Заявка успешно обработана!";
     }
 
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     @Transactional
-    public void approve(String phoneNumber) {
+    public String approve(String phoneNumber) {
         RequestFit requestFit = requestFitRepo.findByPhoneNumber(phoneNumber);
         if (requestFit == null) {
-            throw new EntityNotFoundException(HttpStatus.NOT_FOUND,
-                    "Заявка с указанным номером телефона не найдена");
+            throw new ResourceNotFoundException("Заявка с указанным номером телефона: "
+                    + phoneNumber + " не найдена!");
         }
         Client client = clientRepo.findByPhoneNumber(phoneNumber);
 
         if (client == null) {
-            throw new EntityNotFoundException(HttpStatus.NOT_FOUND,
-                    "Клиент с указанным номером телефона не найден в базе");
+            throw new ResourceNotFoundException("Клиент с указанным номером телефона: "
+                    + phoneNumber + " не зарегистрирован!");
         }
 
         requestFit.setApproved(true);
@@ -134,15 +144,17 @@ public class RequestFitService {
 
         sendMessage(client.getEmail(), "Одобрено", "Одобрено! Ваша заявка ждет оплаты");
 
+        return "Заявка успешно обработана!";
+
     }
 
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     @Transactional
-    public void subscriptionPayment(BigDecimal money, String email) {
+    public String subscriptionPayment(BigDecimal money, String email) {
         Client client = clientRepo.findByEmail(email);
         if (client == null) {
-            throw new EntityNotFoundException(HttpStatus.NOT_FOUND,
-                    "Клиент с указанной электронной почтой не найден в базе");
+            throw new ResourceNotFoundException("Клиент с указанной электронной почтой: "
+                    + email + " не зарегистрирован!");
         }
         client.setBalance(client.getBalance().subtract(money));
         clientRepo.save(client);
@@ -150,6 +162,8 @@ public class RequestFitService {
         Accountant accountant = accountantService.findByName(NAME);
         accountant.setBalance(accountant.getBalance().add(money));
         accountantService.save(accountant);
+
+        return "Оплата прошла успешно!";
     }
 
     @PreAuthorize("hasRole('ROLE_ADMIN')")
